% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prefix_suffix_colnames.R
\name{add_prefix_suffix_column}
\alias{add_prefix_suffix_column}
\title{add_prefix_suffix_column}
\usage{
add_prefix_suffix_column(df, prefix = "", suffix = "",
  skip_cols = "path_patient_id")
}
\arguments{
\item{skip_cols}{}
}
\description{
add_prefix_suffix_column
}
