% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColMeta.R
\name{addColMeta}
\alias{addColMeta}
\title{Add Annotations to Col Meta}
\usage{
addColMeta(adat, apt.data)
}
\arguments{
\item{adat}{A \code{soma_adat} object with intact attributes,
specifically the \code{Col.Meta} element.}

\item{apt.data}{A modified "annotations"
\code{tibble} object to add to attributes.}
}
\value{
A \code{soma_adat} object with new attributes and \code{Col.Meta}
corresponding to the \emph{additional} columns of \code{apt.data}
not present in the existing ADAT.
Rows \emph{not} present in \code{apt.data} will contain an \code{NA} in
the resulting attributes.
}
\description{
Uses modified analyte info of annotations
generated using \code{\link[=getAnalyteInfo]{getAnalyteInfo()}} and adds it to the
\code{Col.Meta} attribute of a \code{soma_adat}
(see \code{attributes(adat)$Col.Meta}).
The column names of the ADAT are \emph{not} conserved.
This could be considered the inverse of \code{\link[=getAnalyteInfo]{getAnalyteInfo()}}.
}
\examples{
adat <- sim_test_data
ad <- SomaDataIO::getAnalyteInfo(adat)
ad

# Adds 2 new 'random' variables:
ad$random1 <- round(rnorm(nrow(ad)), 2)
ad$random2 <- round(rnorm(nrow(ad)), 2)
new <- addColMeta(adat, ad)
SomaDataIO::getAnalyteInfo(new) |> select(AptName, SeqId, starts_with("random"))
}
\seealso{
\code{\link[dplyr:mutate-joins]{dplyr::left_join()}}, \code{\link[=getAnalyteInfo]{getAnalyteInfo()}}, \code{\link[=getSeqId]{getSeqId()}}
}
\author{
Stu Field, Eduardo Tabacman
}
