% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTargetInfo.R
\name{addTargetInfo}
\alias{addTargetInfo}
\title{Add Target Information to Data Table}
\usage{
addTargetInfo(df, apt.data, add.fields = NULL)
}
\arguments{
\item{df}{A data frame object (e.g. the \code{stat.table} of univariate
KS-tests) with \code{Aptamer} row names.}

\item{apt.data}{A \code{tibble} containing analyte specific annotation data
and/or SomaScan run data (e.g. \code{AptName}, \code{Dilution}, \code{Target},
\code{SomaId}, \code{ColCheck}, \code{CalReference}, etc.), see \code{\link[=getAnalyteInfo]{getAnalyteInfo()}}.}

\item{add.fields}{Additional fields to include from the \code{apt.data} object.}
}
\value{
A data frame with original columns of \code{df}, plus these
additional fields:
\itemize{
\item "Target"
\item "EntrezGeneSymbol"
\item "SomaId"
\item "UniProt"
\item any fields specified by \code{add.fields}
}

merged by \code{AptName} as leading columns in the returned data frame.
}
\description{
This function takes a tibble of existing column meta data,
created with \code{\link[=getAnalyteInfo]{getAnalyteInfo()}}, and adds the following:
\itemize{
\item \code{Target}
\item \code{EntrezGeneSymbol}
\item \code{SomaId}
\item \code{UniProt}
}
to a data frame (e.g. a \code{stat.table} of statistical results) which
contains \code{AptName}s as rownames. The rownames will be used to match
\code{SeqIds} when joining the two tables, see \code{\link[=left_join]{left_join()}}.
}
\examples{
anno <- SomaDataIO::getAnalyteInfo(sim_test_data)

# Generate dummy `df` object; e.g. `stat.table`
df <- withr::with_seed(1, {
  data.frame(
    statistic = rnorm(10),
    p.value   = runif(10),
    rank      = 1:10,
    row.names = sample(anno$AptName, 10)    # important; merge by field
  )
})
df

# AptNames of interest (10)
dplyr::filter(anno, AptName \%in\% rownames(df))

# Join tables with Annotations (`anno`)
addTargetInfo(df, apt.data = anno)
}
\seealso{
\code{\link[=getAnalyteInfo]{getAnalyteInfo()}}, \code{\link[=read_adat]{read_adat()}}, \code{\link[=left_join]{left_join()}}
}
\author{
Stu Field
}
