% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{anno}
\alias{anno}
\alias{get_anno}
\alias{grep_anno}
\alias{seq_lookup}
\alias{seqify}
\title{Protein Annotations}
\usage{
get_anno()

grep_anno(pattern)

seq_lookup(seq, tbl = NULL)

seqify(seq)
}
\arguments{
\item{pattern}{\code{character(1)}. Regular expression pattern to match.}

\item{seq}{\code{character(n)}. A vector of \code{SeqIds}, or
strings \emph{containing} \code{SeqId}s.}

\item{tbl}{A \code{tibble} containing annotation data.}
}
\value{
\code{\link[=get_anno]{get_anno()}}: a \code{tibble} of protein annotation information.

\code{\link[=grep_anno]{grep_anno()}}: a \code{tibble} of a subset of
protein information (annotations) corresponding
to matching rows of in \code{pattern}.

\code{\link[=seq_lookup]{seq_lookup()}}: a \code{tibble}, a subset of \code{tbl},
corresponding to the rows whose \code{SeqIds} match the values in \code{seq}.

\code{\link[=seqify]{seqify()}}: a \code{splyr_seq} class object.
}
\description{
A group of functions that allow for the simple
search and access of protein annotation information
based on an internal lookup table.
}
\details{
Historical analytes may have been dropped from the menu
may be absent from the internal lookup table,
resulting in \code{NA} for that row. These
analytes can be retrieved by explicitly passing an annotations
table corresponding to the original data prior to
the menu change. The easiest way to generate a
"time-capsuled" annotations table:
1. via \code{attr(data, "Col.Meta")}
}
\section{Functions}{
\itemize{
\item \code{get_anno()}: a convenient wrapper to easily retrieve the
lookup table (\code{tibble}) of proteomic
annotations keyed on \code{SeqId}.

\item \code{grep_anno()}: pattern matches via regular expression of
the internal "annotations" table (\code{tibble}).
Matches are performed on \emph{any} (\code{regexpr = "|"}) of the following:
\itemize{
\item \code{EntrezGeneSymbol}
\item \code{TargetFullName}
\item \code{Target}
\item \code{List}
\item \code{SeqId}
}

\item \code{seq_lookup()}: conveniently looks up annotation keyed on
a vector of \code{SeqId}s.
The internal lookup dictionary is matched on values in the \code{SeqId}
column of annotations, which accesses a static internal object.
Alternatively, the user can explicitly pass an annotations table of
of their choice.

\item \code{seqify()}: converts to \code{splyr_seq} object, primarily to
dispatch the S3 print method during interactive use.

}}
\examples{
get_anno()

grep_anno("^MMP")     # match EntrezGeneSymbols
grep_anno("^29")      # match SeqId
grep_anno("^black")   # Blacklisted

svec <- c("seq.2981.9", "seq.5073.30", "seq.4429.51", "seq.2447.7")
svec

seq_lookup(svec)

# print method for class `splyr_seq`
seqify(svec) |> class()

# works with seq.xxxx.xx format
seqify(svec)

# also works with pure SeqIds
vec <- sub("\\\\.", "-", sub("^seq\\\\.", "", svec))
vec

seqify(vec)

}
\author{
Stu Field
}
