% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-date.R
\name{BinDates}
\alias{BinDates}
\title{Date Binning}
\usage{
BinDates(date, by = "month", each = 1)
}
\arguments{
\item{date}{Date to be parsed}

\item{by}{Level at which to aggregate, such as \code{hour} or \code{day}; defaults
to \code{month}}

\item{each}{Number of \code{by} periods to combine}
}
\description{
This function takes a vector of dates and rolls them up to the specified time
unit, setting all smaller increments to 0 or 1 as applicable.
}
