% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-cv.R
\name{calc_cv}
\alias{calc_cv}
\title{Calculate CV Decomposition}
\usage{
calc_cv(x)
}
\arguments{
\item{x}{A list of numeric vectors, each representing a separate assay run,
i.e. a technical replicate. The entries of the vectors should be replicates
of intra-run assay variation. \strong{Important!}: be sure to remove any \code{NAs},
if present, beforehand.}
}
\value{
A named vector of the CV decomposition for the replicates.
}
\description{
Calculate the appropriate within and between assay
run variance decomposition prior to calculating
the \%CV for a reproducibility study assay run.
}
\examples{
# Example 1
example <- list(A = c(5785.1, 5005.6, 5686.3, 5990.8,
                      5235.4, 5340.6, 5272.6, 5905.2),
                B = c(5708.0, 5300.7, 5339.8, 5393.0,
                      5762.0, 5553.4, 6081.4, 5473.5),
                C = c(5409.4, 5353.6, 5398.0, 5631.2,
                      5646.1, 5073.4, 5879.2, 5617.5))
example

calc_cv(example)

# Example 2 (dummy `simdata` example)
# look at how the intra, inter, and total variance.
# We assume `simdata` contains *only* replicate samples
# and assume the `SiteId` represents assay "runs".
# We can then `lapply()` over each feature, calculating
# the `CVdecomp` for each and store to a data frame
split_on <- simdata$SiteId

# split data
site_split <- feature_matrix(simdata) |> data.frame() |>
  lapply(split, f = split_on)

# iterate over list/analytes; each contains multiple (4x) "runs"
coeffVar <- lapply(site_split, calc_cv)
coeffVar <- do.call(rbind, coeffVar) |> data.frame()
}
\seealso{
\code{\link[helpr:calc_ss]{helpr::calc_ss()}}
}
\author{
Stu Field
}
