% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-summ-tbl.R
\name{create_summ_tbl}
\alias{create_summ_tbl}
\title{Create A Summary Table}
\usage{
create_summ_tbl(
  data,
  group_var,
  .funs = c("min", "median", "mean", "sd", "mad", "max")
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} object containing data for summary.}

\item{group_var}{\code{character(1)}. An unquoted (or quoted) string
containing the indices to group the statistics,
e.g. \code{Group}. If missing, ungrouped statistics are returned.}

\item{.funs}{\code{character(n)}. String(s) of the functions used to
summarize the data. Each function \emph{must} take a vector of data
as input and return a summary scalar, e.g. \code{\link[=mean]{mean()}}.}
}
\value{
A \code{tibble} object with rows (features) and columns as the
summary statistics listed in \code{.funs}.
}
\description{
Create a summary table of all the feature data based on a
grouping variable of meta data. Defaults summary statistics include:
\itemize{
\item \code{min}
\item \code{median}
\item \code{mean}
\item \code{sd}  (standard deviation)
\item \code{mad} (median absolute deviation)
\item \code{max}
}
}
\examples{
create_summ_tbl(sim_adat)
create_summ_tbl(sim_adat, gender)

# Arbitrary 3 groupings
sim_adat$group <- sample(1:3, nrow(sim_adat), replace = TRUE)
create_summ_tbl(sim_adat, "group")
}
\seealso{
\code{\link[=mad]{mad()}}, \code{\link[=median]{median()}}, \code{\link[=sd]{sd()}}, \code{\link[=min]{min()}}, \code{\link[=max]{max()}}
}
\author{
Stu Field
}
