% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLogRatioAdat.R
\name{createLogRatioAdat}
\alias{createLogRatioAdat}
\title{Calculate log-Ratios by TimePoint}
\usage{
createLogRatioAdat(adat, subject.field, time.field, rm.baseline = FALSE)
}
\arguments{
\item{adat}{A \code{soma_adat} or data frame object containing feature
data, subject IDs (PIDs), and time point information.}

\item{subject.field}{An unquoted string of the subject ID field, containing
identifiers for the patients/subjects with multiple measurements.}

\item{time.field}{An unquoted string of the time point field. Must have at
least 2 levels. Particular attention should be made to ensure that the
baseline sample name can be sorted in a meaningful way, i.e.
\code{sort(unique(adat$time.field))} generates the order you expect.}

\item{rm.baseline}{Logical. Should the baseline samples be removed in
the returned data frame?}
}
\value{
Returns a data frame of the log2-ratios vs. baseline at each time
point for each individual in the data set. The result is sorted by subject
and then by time.
}
\description{
Create a data frame (\code{soma_adat}) with entries corresponding
to the log2-ratios by individual \code{subject.id} for each of the
time points provided in the \code{time} argument.
}
\details{
If \code{rm.baseline = FALSE}, the baseline samples are retained for
consistency of the meta data, but values will be zero. This
particularly makes sense for plotting where there baseline
"zeros" might be desired if \code{TRUE}, the number of data frame
rows will correspond to \code{nrow(adat) - n.subjects}, with the baseline
sample being removed for each subject.
}
\examples{
sim_test_data$time <- rep(1:4, 25)        # random time points
sim_test_data$pid  <- rep(1:25, each = 4)  # assign random subjects
a1 <- createLogRatioAdat(sim_test_data, pid, time)    # keep all
dim(a1)
a2 <- createLogRatioAdat(sim_test_data, pid, time, TRUE)  # remove baseline
dim(a2)
}
\seealso{
\code{\link[=log2]{log2()}}
}
\author{
Stu Field
}
