% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummaryTable.R
\name{createSummaryTable}
\alias{createSummaryTable}
\title{Create A Summary Table}
\usage{
createSummaryTable(
  adat,
  group.var,
  .funs = c("min", "median", "mean", "sd", "mad", "max")
)
}
\arguments{
\item{adat}{A \code{soma_adat} or \code{data.frame} object containing RFU data.}

\item{group.var}{An unquoted (or quoted) string containing the indices
to group the statistics, e.g. \code{SampleGroup}. If missing, ungrouped
statistics are returned.}

\item{.funs}{Character. Strings of the functions used to
summarize the data. Each function \emph{must} take a vector of data as input and
return a scalar summary.}
}
\value{
A \code{tibble} object with \code{AptNames} for rows and columns as the
summary statistics listed in \code{.funs}.
}
\description{
Create a summary table of all the feature/SOMAmer data based on a
grouping variable of meta data. Defaults summary statistics include:
\itemize{
\item \code{min}
\item \code{median}
\item \code{mean}
\item \code{sd}  (standard deviation)
\item \code{mad} (median absolute deviation)
\item \code{max}
}
}
\examples{
createSummaryTable(sim_test_data)
createSummaryTable(sim_test_data, gender)

# Arbitrary 3 groupings
sim_test_data$Group <- sample(1:3, nrow(sim_test_data), replace = TRUE)
createSummaryTable(sim_test_data, "Group")
}
\seealso{
\code{\link[=mad]{mad()}}, \code{\link[=median]{median()}}, \code{\link[=sd]{sd()}}, \code{\link[=min]{min()}}, \code{\link[=max]{max()}}
}
\author{
Stu Field
}
