% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-text.R
\name{CurrencyToNumeric}
\alias{CurrencyToNumeric}
\title{Currency String To Numeric}
\usage{
CurrencyToNumeric(data, fields = NA, debug = FALSE)
}
\arguments{
\item{data}{Data frame or character vector to be manipulated}

\item{fields}{Character vector of fields; not necessary if passing a vector}

\item{debug}{If \code{TRUE}, print debug output}
}
\description{
This function takes a vector or data frame of currency values in character
format, strips out comma delimiters and currency symbols, and converts the
result to numeric.
}
\examples{
cash <- c("10,203¢", "$340", "€8,111.10", "£23,040", "¥8300")
CurrencyToNumeric(cash)

cash.frame <- data.frame(c1 = cash, c2 = cash, c3 = cash, stringsAsFactors = FALSE)

CurrencyToNumeric(cash.frame)
# Note that this unlists the data because no fields were included

CurrencyToNumeric(cash.frame, fields=1:2)
CurrencyToNumeric(cash.frame, fields=c("c2", "c3"))
}

