% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct-at.R
\name{distinct_at}
\alias{distinct_at}
\title{Select Distinct Rows by a Selection of Variables}
\usage{
distinct_at(x, ...)
}
\arguments{
\item{x}{A data frame or tibble object.}

\item{...}{Additional arguments to \code{\link[=select]{select()}}.}
}
\value{
A data frame or tibble object, same class as \code{x}.
}
\description{
Extract distinct rows. Additional arguments allow
specification of variables to check for unique identity.
}
\examples{
withr::with_seed(111, {
  df <- tibble::tibble(
    a_ie = sample(letters[1:3L], 20, replace = TRUE),
    b_ie = sample(1:3, 20, replace = TRUE),
    var  = runif(20)
  )
})
df

distinctAt(df, a_ie, b_ie)

# or with dplyr helper verbs ...
distinctAt(df, ends_with("ie"))
}
\author{
Stu Field
}
