% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct-rows.R
\name{distinct_rows}
\alias{distinct_rows}
\title{Select Distinct Rows}
\usage{
distinct_rows(x, ...)
}
\arguments{
\item{x}{A data frame or tibble object.}

\item{...}{Additional arguments to \code{\link[dplyr:select]{dplyr::select()}}.}
}
\value{
A data frame or tibble object, same class as \code{x}.
}
\description{
Extract distinct rows by selection variables. Additional
arguments allow specification of variables that are
combined and checked for uniqueness.
}
\examples{
withr::with_seed(111, {
  df <- tibble::tibble(
    a_ie = sample(letters[1:3L], 20, replace = TRUE),
    b_ie = sample(1:3, 20, replace = TRUE),
    var  = runif(20)
  )
})
df

distinct_rows(df, a_ie, b_ie)

# or with tidyselect-helpers
distinct_rows(df, ends_with("ie"))
}
\author{
Stu Field
}
