% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_columns.R
\name{expect_columns}
\alias{expect_columns}
\title{expect_columns}
\usage{
expect_columns(trk, columns)
}
\arguments{
\item{trk}{a df}

\item{columns}{columns we epect to have in the df}
}
\description{
check if expected columns are in the df,
else fail with a descriptive message.
}
