% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-text.R
\name{ExtractCurrency}
\alias{ExtractCurrency}
\title{Extract Currency}
\usage{
ExtractCurrency(string, debug = FALSE)
}
\arguments{
\item{string}{String to search for dollar amounts}

\item{debug}{If \code{TRUE}, print debug output}
}
\description{
This function takes a text string and extracts the dollar amount identified as numeric,
allowing the suffixes K (thousand), M (million), B (billion), and T (trillion)
}
\examples{
ExtractCurrency('$11')

ExtractCurrency('$100 total')

ExtractCurrency('Receipt of $11.57')

ExtractCurrency('$10K-$100K', debug = TRUE) # Returns 10,000 (first match)

strings <- c('Nearly $23K', 'A $200M Powerball', 'A $3B endowment', 'GDP last year of $17.1T (USD)')
ExtractCurrency(strings, debug = TRUE)
}
