% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-matrix.R
\name{feature_matrix}
\alias{feature_matrix}
\title{Select the Feature Matrix}
\usage{
feature_matrix(data, feat = NULL)
}
\arguments{
\item{data}{A \code{data.frame} object containing
numeric feature data.}

\item{feat}{\code{charcter(n)}. A vector of column
names corresponding to the features.}
}
\value{
A numeric \code{data.matrix} object containing
only the data matrix of values for the features.
}
\description{
Selects the feature (variables) data from
a \code{data.frame} object and returns a data matrix
of only the feature (usually numeric) data.
Particularly useful in selecting prior to
\code{\link[stats:predict]{stats::predict()}} methods that require
a named, numeric data matrix as input.
}
\examples{
dim(sim_adat)
class(sim_adat)

feature_matrix(sim_adat) |> dim()
feature_matrix(sim_adat) |> class()
}
\seealso{
\code{\link[=data.matrix]{data.matrix()}}
}
\author{
Stu Field
}
