% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{fix_names}
\alias{fix_names}
\title{Fix names of the data.frame}
\usage{
fix_names(x, char = "-")
}
\arguments{
\item{x}{vector of sample names to be fixed}

\item{char}{The character seperator desired}
}
\description{
Fix names of the data.frame
}
