% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-anno.R
\name{get_anno}
\alias{get_anno}
\title{Get Database Annotations}
\usage{
get_anno()
}
\value{
A tibble.
}
\description{
A convenient wrapper to easily retrieve the
lookup table (tibble) of proteomic
annotations keyed on \code{SeqId}.
}
\examples{
get_anno()
}
\author{
Stu Field
}
