% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotations.R
\name{getAnnotations}
\alias{getAnnotations}
\alias{getAnnotationsAPI}
\title{Get Graph Database Annotations}
\usage{
getAnnotations(api = FALSE, path = "annotations_graph/api")

getAnnotationsAPI(path = "annotations_graph/api")
}
\arguments{
\item{api}{Logical. Should the API endpoint be queried, if
available, for live updated annotation information?}

\item{path}{The umbrella API endpoint address. Ignored if \code{api = FALSE}.}
}
\value{
A tibble of the corresponding column meta data available in
the Annotations Graph DB.
}
\description{
Generate a lookup table (tibble) of annotations
from the Graph Database keyed on \code{SeqId}.
By default, an internal static object
corresponding to analyte annotations is returned.
Alternatively, a request is sent to the
internal Graph DB API endpoint to pull a
live/dynamic annotations table via a call
to \code{\link[=getAnnotationsAPI]{getAnnotationsAPI()}}.
}
\section{Functions}{
\itemize{
\item \code{getAnnotationsAPI()}: Similar to \code{\link[=getAnnotations]{getAnnotations()}} but requests
live dynamic annotations from the Graph DB API.

}}
\examples{
getAnnotations()
}
\seealso{
Other Annotations: 
\code{\link{lookupAnnotations}()},
\code{\link{seqLookup}()}
}
\author{
Stu Field
}
\concept{Annotations}
