% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-date.R
\name{GetFYDate}
\alias{GetFYDate}
\alias{GetFYEnd}
\alias{GetFYStart}
\title{Fiscal Date Generation}
\usage{
GetFYStart(date, start.month = 7)

GetFYEnd(date, start.month = 7)
}
\arguments{
\item{date}{Dates to be parsed}

\item{start.month}{Starting month number of the fiscal year; defaults to 7 (July)}
}
\description{
These functions take a vector of dates and return the fiscal year start or end dates.
}
\examples{
(dts <- ToDate(paste("2012", 1:12, 1, sep="-")))

GetFYStart(dts)
GetFYEnd(dts)

GetFYStart(dts, start.month = 1)
GetFYEnd(dts, start.month = 1)
# Properly handles calendar years (i.e. no conversion)

GetFYEnd(dts, start.month = 3)
# Properly accounts for leap years
}

