% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced-functions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{left-hand side}

\item{y}{right-hand side}
}
\description{
This function returns the complement of \code{\%in\%} or \code{match()}.
}
\examples{
# Compare !(\%in\%) to \%nin\%
!(1:3 \%in\% 2:4)
1:3 \%nin\% 2:4
}

