% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-labels.R
\name{group_labels}
\alias{group_labels}
\title{Return Grouping Labels}
\usage{
group_labels(x)
}
\arguments{
\item{x}{A data frame, \code{tibble} or \verb{grouped tibble}.}
}
\description{
Returns a \code{tibble} of the grouping labels (usually class labels)
corresponding a \code{grouped_df} class object.
}
\examples{
df <- dplyr::group_by(mtcars, cyl)
class(df)
dplyr::group_data(df)
dplyr::group_vars(df)

# 1 variable
group_labels(df)

# 2 variables
df <- dplyr::group_by(mtcars, cyl, vs)
group_labels(df)
}
\seealso{
\code{\link[dplyr:group_data]{dplyr::group_vars()}}, \code{\link[dplyr:group_data]{dplyr::group_data()}}
}
\author{
Stu Field
}
