% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\alias{impute_median}
\alias{impute_missing}
\title{Impute Missing Values}
\usage{
impute_median(x)

impute_missing(x, distr = c("norm", "gamma", "beta"), seed = 123)
}
\arguments{
\item{x}{A \code{data.frame}, matrix, or numeric vector. Factors or
character strings are also possible but see caveat above.}

\item{distr}{\code{character(1)}. The name of the desired distribution.}

\item{seed}{\code{integer(1)}. Set a seed for reproducibility.}
}
\value{
An object of the same dimension and class as \code{x}
with any \code{NA} (missing) values replaced.
}
\description{
There are two imputation strategies:

Imputing with the median is typically performed for
numeric vectors, however methods also exist for:
\itemize{
\item the \code{factor} method imputes to the most common factor level.
\item the \code{character} method imputes to the most common string
}

Imputing missing values from an estimated distribution is
performed by fitting via Maximum Likelihood and estimating
and parameters, followed by random draws from the appropriate
random generator.
Currently only supports Gamma and Gaussian distributions.
}
\examples{
# data frame method
df <- select(mtcars, mpg, cyl, disp) |> head()
df[cbind(2:3, 2:3)] <- NA_real_
df

impute_median(df)
x <- rnorm(10, mean = 10, sd = 5)
x <- rbeta(10, shape1 = 2, shape2 = 3)
x[c(2, 7)] <- NA_real_

data.frame(
  orig    = x,
  imputed = impute_missing(x)   # gaussian
)
}
\seealso{
\code{\link[=median]{median()}}

Other impute: 
\code{\link{impute_outliers}()},
\code{\link{impute_predictors}()}
}
\author{
Stu Field
}
\concept{impute}
