% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeNAs.R
\name{imputeNAs}
\alias{imputeNAs}
\title{Impute NAs}
\usage{
imputeNAs(x)
}
\arguments{
\item{x}{A \code{data.frame}, matrix, or numeric vector.}
}
\value{
An object of the same dimension and class as \code{x}
with any \code{NA} values replaced with the median of the
corresponding vector.
}
\description{
Imputes any \code{NAs} with the median value.
Must be a numeric data only. \code{factor} or
\code{character} class are not allowed.
}
\examples{
# data frame method
df <- select(simdata, id, class_response, seq.2802.68) |> head()
df[3, 3] <- NA_real_
df

imputeNAs(df)

# vector method
x <- seq(10)
x[c(1, 3, 5)] <- NA_real_
x

median(x, na.rm = TRUE)

imputeNAs(x)

table(imputeNAs(x)) |> sort() |> tail()
}
\seealso{
\code{\link[=median]{median()}}

Other impute: 
\code{\link{impute_outliers}()},
\code{\link{impute_predictors}()}
}
\author{
Stu Field
}
\concept{impute}
