% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan-data.frame.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Checks for NaNs in a data.frame}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Checks for NaNs in a data.frame
}
