% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-list.R
\name{ListExtract}
\alias{ListExtract}
\title{Extract named elements from a nested list}
\usage{
ListExtract(list, name, name.elements = TRUE)
}
\arguments{
\item{list}{List to be reformatted}

\item{name}{Quoted name of sub-list item to extract}

\item{name.elements}{Whether to name the new list elements based on the sub-list
item's name and their order in the original list; defaults to \code{TRUE}}
}
\description{
Given a list of lists, this function goes through each element and copies
the named element from each sub-list into a top-level element of a new list.
}
\examples{

# Create and view a nested list
a <- list(letter = 'a', number = 1)
b <- list(letter = 'b', number = 2)
c <- list(letter = 'c', number = 3)
(nested <- list(first = a, second = b, third = c))

# Extract the letter element
ListExtract(nested, 'letter')

# Extract the number element
ListExtract(nested, 'number')

# Extract the number element into an unnamed list
ListExtract(nested, 'number', FALSE)
}
