% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-anno.R
\name{lookup_anno}
\alias{lookup_anno}
\title{Look Up Annotations by Pattern}
\usage{
lookup_anno(pattern)
}
\arguments{
\item{pattern}{Character. Regular expression pattern to match.}
}
\value{
A tibble, a subset of Annotations from the Graph DB
corresponding to matching rows of the desired \code{pattern}.
}
\description{
Searches the internal "annotations" table (\code{tibble}) from
regular expression matches according to \code{pattern}.
Pattern matching is performed \emph{any} (\code{regexpr = "|"}) of the following:
\itemize{
\item \code{EntrezGeneSymbol}
\item \code{TargetFullName}
\item \code{Target}
\item \code{List}
\item \code{SeqId}
}
}
\examples{
lookup_anno("^MMP")     # match EntrezGeneSymbols
lookup_anno("^29")      # match SeqId
lookup_anno("^black")   # Blacklisted
}
\seealso{
Other Annotations: 
\code{\link{seq_lookup}()}
}
\author{
Stu Field
}
\concept{Annotations}
