% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupAnnotations.R
\name{lookupAnnotations}
\alias{lookupAnnotations}
\title{Look Up Aptamer Pattern}
\usage{
lookupAnnotations(pattern)
}
\arguments{
\item{pattern}{Character. Regular expression pattern to match.}
}
\value{
A tibble, a subset of Annotations from the Graph DB
corresponding to matching rows of the desired \code{pattern}.
An empty \code{tibble} if no matches found.
}
\description{
Searches the internal "Annotations" table (\code{tibble}) from
the Graph Database for regular expression matches according to \code{pattern}.
Pattern matching is performed \emph{any} (\code{regexpr = "|"}) of the following:
\itemize{
\item \code{EntrezGeneSymbol}
\item \code{TargetFullName}
\item \code{Target}
\item \code{List}
\item \code{SeqId}
}
}
\examples{
lookupAnnotations("^MMP")     # match EntrezGeneSymbols
lookupAnnotations("^29")      # match SeqId
lookupAnnotations("^black")   # Blacklisted
}
\seealso{
Other Annotations: 
\code{\link{getAnnotations}()},
\code{\link{seqLookup}()}
}
\author{
Stu Field
}
\concept{Annotations}
