% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNumeric.R
\name{makeNumeric}
\alias{makeNumeric}
\alias{makeNumeric.character}
\alias{makeNumeric.factor}
\alias{makeNumeric.integer}
\alias{makeNumeric.logical}
\alias{makeNumeric.list}
\alias{makeNumeric.data.frame}
\title{Convert Table Entries to Numeric}
\usage{
makeNumeric(x, ...)

\method{makeNumeric}{character}(x, ...)

\method{makeNumeric}{factor}(x, ..., coerce.factor = TRUE)

\method{makeNumeric}{integer}(x, ...)

\method{makeNumeric}{logical}(x, ...)

\method{makeNumeric}{list}(x, ..., coerce.factor = TRUE)

\method{makeNumeric}{data.frame}(x, ..., coerce.factor = TRUE)
}
\arguments{
\item{x}{An object for S3 dispatch. Usually a character,
data frame, tibble, or named list containing convertible columns.}

\item{...}{Additional arguments to allow extensibility to S3 methods.}

\item{coerce.factor}{Logical. Should \code{factor} types be
converted to their corresponding numeric?}
}
\value{
An object coerced to a numeric if the rules above can be
followed. A data frame, \code{tibble}, or list returns the same class object
with the function applied to each element (column).
}
\description{
Selectively convert an R object to a numeric vector if
doing so does \emph{not} result in a warning. This is
typically the case when a \code{character} string
contains a symbol that cannot be coerced cleanly. The following rules
apply during conversion:
\itemize{
\item characters converted if no warnings are triggered
\item integers are \emph{never} converted
\item logical vectors \emph{always} converted to 0 / 1
\item factors are \emph{optionally} converted to numeric
}
}
\examples{
tbl <- tibble::tibble(
  id     = 1:20,                            # NO
  chr_id = as.character(1:20),              # YES
  logic  = sample(c(TRUE, FALSE), 20, replace = TRUE), # YES
  fact   = factor(letters[1:20]),           # YES or NO depending on coerce.factor
  num    = rnorm(20),                       # YES
  x      = rep(c("foo", "bar"), each = 10), # NO
  y      = c(as.character(runif(19)), NA),  # YES; only 1 NA, 5\% total
  z      = as.character(rnorm(20))          # YES
)
makeNumeric(tbl)
}
\seealso{
\code{\link[=as.numeric]{as.numeric()}}, \code{\link[=modify]{modify()}}
}
\author{
Stu Field
}
