% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-samples.R
\name{match_samples}
\alias{match_samples}
\title{Match Two Data Frames Based on Sample ID}
\usage{
match_samples(x, y, idcol = "SampleId")
}
\arguments{
\item{x, y}{The two data frames to match samples against.}

\item{idcol}{\code{character(1)}. The name of the column containing
sample identifiers in both \code{x} and \code{y}.}
}
\value{
A named list with \code{x} and \code{y}, matched and ordered
based on \code{idcol}.
}
\description{
Given two data frames, this function matches their rows
based on the intersection of a unique sample identifier column,
\code{idcol}.
The returned data frames are ordered by \code{idcol} and containing
only rows based on the intersection of \code{idcol} values.
}
\examples{
df    <- head(simdata, 25L)
train <- withr::with_seed(1, dplyr::sample_n(df, 15L))
new   <- withr::with_seed(2, dplyr::sample_n(df, 15L))
intersect(train$SampleId, new$SampleId)  # overlapping IDs exist

dfs <- match_samples(train, new)
dfs
}
