% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{mclapply2}
\alias{mclapply2}
\title{Runs multicore lapply with progress indicator and transformation to}
\usage{
mclapply2(X, FUN, ...)
}
\arguments{
\item{X}{a list or vector, to loop over}

\item{FUN}{a function}

\item{...}{other arguments}
}
\description{
mclapply2
}
\examples{
\dontrun{
mclapply2(X = X, FUN = FUN, ... = ...)
mclapply2(X = cmds, FUN = function(x, mc.cores = 5){system(x)})
}
}
