% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotLong.R
\name{pivotLong}
\alias{pivotLong}
\title{Convert Wide & Long Mixed Data to Long Format}
\usage{
pivotLong(data, sub.key = "subject_id", tp.key = "ie", clin.sep = "_ie")
}
\arguments{
\item{data}{A clinical data frame with \emph{both} "wide" and "long"
format containing duplicated clinical information.}

\item{sub.key}{Character. A quoted string of the column name containing
a unique identifier corresponding to each individual.}

\item{tp.key}{Character. A quoted string identifying the variable (column)
name that contains the time point index for the proteomic data. This
variable is \emph{hard-coded} in the form \verb{"IE<x>"} (\code{x} = time-point),
and is over-written during the process.}

\item{clin.sep}{Character. A quoted string containing a
unique separator which exists in a variable name between the
clinical variable and the time point index.}
}
\value{
A data frame object of the same class as \code{data} converted
to long format. The returned object has one row per sample
identifier, only one column for a given clinical covariate, and
the same number of rows as the input \code{data}.
}
\description{
Takes a data frame object which has time series clinical
measurements which are stratified across columns in wide format.
Thus the data exists in both "long" \emph{and} "wide" format and contains
duplicated information.
Clinical variables are typically named in a specific format:
\verb{<clinical_variable><clin.sep><timepoint>}.
The purpose of \code{\link[=pivotLong]{pivotLong()}} is to restructure the data to standard
long format used in SomaLogic proteomic analyses.
}
\examples{
widelong <- tibble::tibble(
  subject_id    = c("a", "a", "a", "b", "b", "b"),
  ie            = c("IE1", "IE2", "IE3", "IE1", "IE2", "IE3"),
  clin_var1_ie1 = c(5L, 5L, 5L, 20L, 20L, 20L),
  clin_var1_ie2 = c(10L, 10L, 10L, 25L, 25L, 25L),
  clin_var1_ie3 = c(15L, 15L, 15L, 30L, 30L, 30L),
  clin_var2_ie1 = c(1.5, 1.5, 1.5, 7.7, 7.7, 7.7),
  clin_var2_ie2 = c(9.9, 9.9, 9.9, 9.2, 9.2, 9.2),
  clin_var2_ie3 = c(8.8, 8.8, 8.8, 1.6, 1.6, 1.6),
  clin_var3_ie1 = c("N", "N", "N", "Y", "Y", "Y"),
  clin_var3_ie3 = c("Y", "Y", "Y", "N", "N", "N"),
  seq.15529.33  = seq(34567, 45678, length.out = 6)
)

piv <- pivotLong(widelong)
piv

# samples (rows) by id are preserved
all.equal(nrow(piv), nrow(widelong))
}
\author{
Stu Field, Jeremy Primus
}
