% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance.R
\name{rebalance}
\alias{rebalance}
\title{Sampling for Class Imbalances}
\usage{
rebalance(data, var, method = c("down", "up"))
}
\arguments{
\item{data}{A data frame or tibble. For example, an individual
"analysis" fold (\code{\link[rsample:as.data.frame.rsplit]{rsample::analysis()}}) from an \code{\link[rsample:vfold_cv]{rsample::vfold_cv()}} call.}

\item{var}{\code{character(1)}. A column name of \code{data} containing class labels
for up-/down-sampling. Assumes a binary variable.
\code{tidyselect} variable selection is also supported (see examples).}

\item{method}{\code{character(1)}. One of "down" or "up" (matched).
Down-sampling \emph{decreases} size of the major class,
whereas up-sampling \emph{increases} (i.e. sampling with
replacement) the size per group of the minor class.}
}
\value{
A sub- or re-sampled data frame that is either smaller or
larger than \code{data} with equal class balance.
}
\description{
Implements up- or down-sampling for analyses
with binary class imbalances.
}
\examples{
xtab <- helpr::cross_tab

# original class distribution
xtab(mtcars, vs)

# down-sampling `vs`
si <- rebalance(mtcars, "vs")    # 'down'
xtab(si, vs)

# up-sampling `vs` & passed variable
var <- "vs"
si <- rebalance(mtcars, var, "up")
xtab(si, vs)

# also supports unquoted strings
si  <- rebalance(mtcars, vs)
xtab(si, vs)

}
\author{
Stu Field
}
