% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactorData.R
\name{refactorData}
\alias{refactorData}
\title{Refactor Ghost Level Meta Data}
\usage{
refactorData(data)
}
\arguments{
\item{data}{A \code{soma_adat} object containing RFU and meta data.}
}
\value{
A \code{soma_adat} object, identical as the
original except the factor levels in factor class columns
have been refactored to remove "ghost levels".
}
\description{
Refactor the meta data fields in a data frame,
removing the "ghost" levels that remain after
subsetting a factored column in a data frame.
}
\examples{
sim_test_data$Sex <- factor(sim_test_data$gender)
sim_test_data$Sex

new <- sim_test_data[sim_test_data$Sex == "M", ]
new$Sex           # ghost levels!
levels(new$Sex)   # ghost levels!

new2 <- refactorData(new)
new2$Sex
levels(new2$Sex)
}
\seealso{
\code{\link[=getMeta]{getMeta()}}, \code{\link[=factor]{factor()}}, \code{\link[=droplevels]{droplevels()}}
}
\author{
Stu Field
}
