% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBufferSamples.R
\name{removeBufferSamples}
\alias{removeBufferSamples}
\title{Remove Buffer Samples From an Object}
\usage{
removeBufferSamples(data)
}
\arguments{
\item{data}{A \code{soma_adat}, \code{data.frame}, or \code{tibble} object.}
}
\value{
An object the same class as \code{data} having
removed rows representing the \code{Buffer} samples,
i.e. \code{SampleType == "Buffer"}.
}
\description{
Removes the buffer samples from a \code{soma_adat}
object assuming the \code{SampleType} field is
present in the object and it contains buffer samples.
Additional fall-back to the \code{SampleId} field
for older ADATs; otherwise a warning is triggered
and no samples are removed. \code{Buffer} is case-insensitive.
}
\examples{
df <- withr::with_seed(101, {
  data.frame(
    PlateId = rep("Set A", 20L),
    SampleType = sample(c("Sample", "QC", "Calibrator", "Buffer"),
                        size = 10L, replace = TRUE),
    clinvar = rnorm(20)
  )
})
table(df$SampleType)

removeBufferSamples(df)$SampleType |> table()
}
