% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-text.R
\name{RenameHeaders}
\alias{RenameHeaders}
\title{Rename Headers}
\usage{
RenameHeaders(dframe, old, new, debug = FALSE)
}
\arguments{
\item{dframe}{Data frame to be manipulated}

\item{old}{Character vector of old header names}

\item{new}{Character vector of new header names, same length as old}

\item{debug}{If TRUE, print debug output}
}
\description{
This function renames a set of data frame headers
}
\examples{
iris <- RenameHeaders(iris, old=c("Sepal.Length", "Sepal.Width"),
                      new=c("Sep.L", "Sep.W"))
}

