% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAnalytes.R
\name{scaleAnalytes}
\alias{scaleAnalytes}
\title{Scale/transform Analyte RFUs}
\usage{
scaleAnalytes(.data, scale_vec)
}
\arguments{
\item{.data}{A \code{soma_adat} class object ... \emph{must} be a \code{soma_adat} for
downstream methods to properly dispatch.}

\item{scale_vec}{A vector of scalars, named by \code{SeqId}, see \code{\link[=getSeqId]{getSeqId()}}.}
}
\description{
Scale analytes by a scalar reference value with \code{SeqId} names to
match with the analytes contained in \code{.data}. Columns without a
corresponding reference value are not modified (with a warning).
}
\examples{
adat <- head(sim_test_data, 3L) |> dplyr::select(id, c(16:18))
ref  <- c("2802-68" = 10.0, "1942-70" = 0.1, "9251-29" = 1.0)
new  <- scaleAnalytes(adat, ref)
new
}
\author{
Stu Field
}
