% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLookup.R
\name{seqLookup}
\alias{seqLookup}
\alias{seqify}
\title{SeqId Look-Up Table}
\usage{
seqLookup(seq, tbl = NULL)

seqify(x)
}
\arguments{
\item{tbl}{A tibble containing annotation data.}

\item{x, seq}{Character. A vector of \code{SeqIds}, or
strings containing \code{SeqId}s.}
}
\value{
For \code{\link[=seqLookup]{seqLookup()}}, a tibble, a subset of \code{tbl},
corresponding to the rows whose \code{SeqIds} match the values in \code{seq}.

For \code{seqify()}, an object of class \code{seq}.
}
\description{
Conveniently look up "annotation" information such as:
\itemize{
\item \code{EntrezGeneSymbol}
\item \code{Target}
\item \code{Dilution}
\item \code{UniProt}
}

keyed on a vector of \code{SeqId}s.
The lookup dictionary is matched on values in the \code{SeqId}
column of annotations, which accesses a static internal object.
Alternatively, the user can explicitly pass an annotations table of
of their choice.
The \code{\link[=seqify]{seqify()}} function adds a \code{seq} class attribute allowing
a convenient S3 print dispatch for interactive use.
}
\details{
If historical analytes have been dropped from the menu
they will be absent from the default lookup table,
resulting in \code{NA} for that row. But not all is lost! Dropped
analytes can be retrieved by explicitly passing an annotations
table corresponding to the original data prior to
the menu change (see \code{Examples}). The easiest way to generate a
"time-capsuled" annotations table:
1. via \code{attr(data, "Col.Meta")}
}
\section{Functions}{
\itemize{
\item \code{seqify()}: Convert to \code{seq} object.

}}
\examples{
svec <- c("seq.2981.9", "seq.5073.30", "seq.4429.51", "seq.2447.7")
svec

seqLookup(svec)

# can pass 'known' set of annotations
#   note: the NAs are now replaced with updated info
seqLookup(svec, tbl = splyr:::apt_data)

# print method for class `seq`
seqify(svec) |> class()

# works with seq.xxxx.xx format
seqify(svec)

# also works with pure SeqIds
vec <- sub("\\\\.", "-", sub("^seq\\\\.", "", svec))
vec

seqify(vec)
}
\seealso{
\code{\link[=getAnnotations]{getAnnotations()}}, \code{\link[=left_join]{left_join()}}

Other Annotations: 
\code{\link{lookupAnnotations}()}
}
\author{
Stu Field
}
\concept{Annotations}
