% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated Data Object}
\format{
The format of \code{simdata} (100 x 47) is:
\describe{
\item{id}{A sample \code{id} number.}
\item{class_response}{The classification response variable.}
\item{reg_response}{The regression response variable.}
\item{time}{The time-to-event variable for survival problems.}
\item{status}{The event status variable for survival problems.}
\item{SampleId}{Identical to the \code{id} column; for \code{POC} tools to run.}
\item{SlideId}{Randomly generated numeric, used to create rownames.}
\item{Subarray}{All \code{1}s. Used to create rownames.}
\item{SiteId}{Locations from \code{LOTR}; for \code{POC} tools to run.}
\item{age}{A randomly assigned age variable: 18 - 80.}
\item{gender}{A randomly assigned gender variable: "F" or "M".}
\item{HybControlNormScale}{Random continuous data in \verb{[0.4, 2.5]}.}
\item{NormScale_40}{Random continuous data in \verb{[0.4, 2.5]}.}
\item{NormScale_0_005}{Random continuous data in \verb{[0.4, 2.5]}.}
\item{NormScale_1}{Random continuous data in \verb{[0.4, 2.5]}.}
}

General features of the data (see \code{attributes(simdata)}):
\itemize{
\item 5 significant classification features
\item 5 significant regression features
\item 5 significant survival features
\item 25 noise features
\item 4 sex/gender features (with noise features)
\item \code{PGAM} (\code{3896-5} for sample handling)
}
}
\usage{
simdata
}
\description{
Simulates proteomic data by randomly generating data for various
Bioinformatics related analyses and unit tests. Feature names are randomly
generated and have the \code{seq.xxxx.xx} format so that downstream tools will
function seamlessly. Model types this data can support:
\enumerate{
\item Binary classification
\item Continuous regression
\item Survival
}
}
\examples{
simdata
}
\author{
Stu Field
}
\keyword{datasets}
