% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaRecipe.R
\name{somaRecipe}
\alias{somaRecipe}
\alias{somaBake}
\alias{is.baked}
\alias{convertRecipe}
\title{Pre-processing SomaScan Data}
\usage{
somaRecipe(
  data,
  log10 = TRUE,
  center = TRUE,
  scale = TRUE,
  bridge_ref = NULL,
  ...
)

somaBake(x, data)

is.baked(data)

convertRecipe(object)
}
\arguments{
\item{data}{A \code{soma_adat} class object to be pre-process according
to \code{x}.}

\item{log10}{Logical. Should \code{data} be log10-transformed.}

\item{center}{Logical. Indicating whether the variables
should be shifted to be zero centered (\eqn{\mu = 0}).}

\item{scale}{Logical. Indicating whether the variables
should be scaled to have unit variance (\eqn{\sigma = 1}).}

\item{bridge_ref}{Numeric. A vector of scale factors to bridge transform
the data \emph{prior} to applying the remaining steps in the recipe.
They are passed to \code{\link[=scale_features]{scale_features()}}, and should be take the same
format as the \code{scale_vec} argument. If \code{NULL} (default), no
bridging is performed.}

\item{...}{Optional arguments of the form \verb{variable = function()}
specifying the function to be applied to the specified column.
Anonymous functions can be used but they should take a vector
input and return a vector of equal length. See Examples.}

\item{x}{A \code{soma_recipe} class object with instructions for pre-processing.}

\item{object}{A \code{recipe} class object from the \code{recipes} package with
the 3 pre-processing steps described above.}
}
\description{
Relative too the \pkg{recipes} package, the \code{\link[recipes:recipe]{recipes::recipe()}} and
\code{\link[recipes:prep]{recipes::prep()}} functions are combined.
}
\details{
The order of recipe steps is \emph{always}:
\enumerate{
\item bridging (if \code{bridge_ref} passed)
\item log10-transform
\item center
\item scale
}
}
\section{Functions}{
\itemize{
\item \code{somaBake()}: executes the recipe instructions defined during
via \code{\link[=somaRecipe]{somaRecipe()}}.

\item \code{is.baked()}: tests for presence of \code{baked} entry in
attributes, indicating that these data have already been baked via
\code{\link[=somaBake]{somaBake()}}.

\item \code{convertRecipe()}: converts an existing recipe object from the
\pkg{recipes} package into a \code{soma_recipe} object. Note that \emph{all}
conversions and not possible and the intended use case is for 3 step
conversion:
\itemize{
\item log10-transform
\item centered
\item scaled
}

}}
\examples{
# create a pre-processing recipe
rcp <- somaRecipe(sim_test_data)

# additional processing of meta data variables
rcp2 <- somaRecipe(sim_test_data,
                   SiteId = factor,
                   reg_response = abs,
                   time = function(x) sqrt(x + pi))

# apply recipe to orig/own data set
new_data <- somaBake(rcp, sim_test_data)

# Logical test
is.baked(new_data)

# converting recipes
library(recipes)
rec <- recipe(Species ~ ., data = iris) |>
  step_log(all_predictors(), base = 10) |>
  step_center(all_predictors()) |>
  step_scale(all_predictors()) |>
  prep()

convertRecipe(rec)
}
