% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_str_df.R
\name{split_str_df}
\alias{split_str_df}
\title{split_str_df}
\usage{
split_str_df(x,
  format = "{{samplename}}_{{index}}_L00{{lane}}_R{{read}}_{{num}}.fastq.gz",
  lst_patterns = list(samplename = "(.*)", index = "([ATGC]*|NoIndex)",
  lane = "([0-9]*)", read = "([0-9]*)", num = "([0-9]*)"),
  strict_format_checking = FALSE, verbose = T)
}
\arguments{
\item{x}{a fastq file}

\item{format}{naming format for the file}

\item{strict_format_checking}{TRUE (error)/FALSE(warning)}
}
\description{
split_str_df
}
