% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splyr-package.R
\docType{package}
\name{splyr-package}
\alias{splyr}
\alias{splyr-package}
\title{splyr: Manipulate and Wrangle Data}
\description{
The splyr package contains the general functions necessary to manipulate internal R representations of `*.adat` files and wrangle those data into convenient form for analysis.
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
\keyword{package}
