% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_mat_df.R
\name{to_mat}
\alias{to_mat}
\title{convert df to mat}
\usage{
to_mat(x, rowname_var = 1)
}
\arguments{
\item{x}{data.frame}

\item{rowname_var}{column ID to use as rowname, default 1}
}
\description{
convert df to mat
}
