% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-text.R
\name{ToDateCleaner}
\alias{ToDateCleaner}
\title{Invalid date string parsing}
\usage{
ToDateCleaner(
  datestring,
  method = "ymd",
  fallback_mo = 1,
  fallback_day = 1,
  debug = FALSE
)
}
\arguments{
\item{datestring}{Numeric string of length 1-8 to interpret}

\item{method}{Position of year, month, day values within the string. Currently ONLY \code{ymd} is supported!}

\item{fallback_mo}{Month number to insert, if necessary}

\item{fallback_day}{Day number to insert, if necessary}

\item{debug}{If \code{TRUE}, print debug output}
}
\description{
This function is a wrapper for ToDate that takes a character string of integers of length
4 to 8 and attempts to parse out a valid date.
}
\examples{
weird_dates <- c('20200101', '2020', '20201', '202001', '20200000'
, '2020001', '20200001', '20200199')

# ToDate returns errors
# But ToDateCleaner attempts to fix the invalid date strings
ToDateCleaner(weird_dates)
}
