% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-recipe-params.R
\name{get_recipe_params}
\alias{get_recipe_params}
\title{Get Parameters From A Recipe}
\usage{
get_recipe_params(recipe, param)
}
\arguments{
\item{recipe}{Prepped \code{recipe} or \code{soma_recipe} class object.}

\item{param}{Character, possible values are 'center' and 'scale'.}
}
\value{
A named numeric vector with means (\code{step = 'center'}) or standard
deviations (\verb{step = scale'}). Names correspond to recipe predictors.
}
\description{
Get mean and standard deviations used to center/scale data from a prepped
recipe.
}
\examples{
test <- sim_test_data
apts <- globalr:::getAnalytes(test)
rec <-  recipes::recipe(~ ., data = dplyr::select(test, dplyr::all_of(apts))) |>
 recipes::step_log(recipes::all_predictors(), base = 10) |>
 recipes::step_center(recipes::all_predictors()) |>
 recipes::step_scale(recipes::all_predictors()) |>
 recipes::prep(training = test)
get_recipe_params(rec, "scale")
get_recipe_params(rec, "center")

soma_rec <- somaRecipe(test)
get_recipe_params(soma_rec, "scale")
get_recipe_params(rec, "center")
}
\author{
Amy Zhang
}
