% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced-functions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{left-hand side}

\item{y}{right-hand side}
}
\description{
This function returns the complement of %in% or match()
}

