% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-outliers.R
\name{impute_outliers}
\alias{impute_outliers}
\title{Impute Outlier Values}
\usage{
impute_outliers(x, n_sigma = 3)
}
\arguments{
\item{x}{A vector of values, approximating a Gaussian distribution
and (possibly) containing outliers.}

\item{n_sigma}{\code{numeric(1)}. The the number of standard
deviations from the mean a \code{n_sigma} threshold for outliers.
Ignored if \code{type = "nonparametric"}.}
}
\value{
A vector approximating a Gaussian
distribution with outliers imputed back
to the robust Gaussian fit.
}
\description{
Given a numeric vector, this function imputes outlier values,
defined as \eqn{3 \times \sigma} from the mean, back to a
robustly calculated Gaussian. Gaussian parameters (\eqn{\mu + \sigma})
are robustly calculated.
}
\details{
The maximum value, i.e. the \eqn{100^{th}} percentile, is
pushed back to the \eqn{3 \times \sigma} value of the Gaussian.
}
\examples{
# Gaussian with 4 outliers (2hi, 2lo)
vec  <- withr::with_seed(101, c(2, 2.5, rnorm(26, 15, 2), 25, 25.9))
pars <- attributes(helpr::get_outliers(vec, n_sigma = 3, type = "para"))
pars

impute_outliers(vec)

# Plot what is happening:
#   original `vec` and critical 3*sd cutoffs in `red`
library(ggplot2)
range <- seq(min(vec), max(vec), length.out = 100)
fit_data <- data.frame(
  x = range,
  y = stats::pnorm(range, mean = pars$mu, sd = pars$sigma)
)
ggplot(data.frame(x = vec), aes(x = x)) +
  stat_ecdf(color = "blue") +
  stat_ecdf(data = data.frame(x = impute_outliers(vec)),
            aes(x = x), color = "purple") +
  geom_line(data = fit_data, aes(x = x, y = y),
            colour = "black", linetype = "longdash") +
  geom_vline(xintercept = pars$crit, colour = "red", linetype = "dashed") +
  ggtitle("Outlier Cutoffs (3*sigma) in Red")

}
\seealso{
\code{\link[=pnorm]{pnorm()}}, \code{\link[=qnorm]{qnorm()}}, \code{\link[helpr:get_outliers]{helpr::get_outliers()}}

Other impute: 
\code{\link{impute}},
\code{\link{impute_predictors}()}
}
\author{
Stu Field
}
\concept{impute}
