% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-predictors.R
\name{impute_predictors}
\alias{impute_predictors}
\title{Impute Predictor Data}
\usage{
impute_predictors(data, extrm_vals)
}
\arguments{
\item{data}{A \code{data.frame}, or \code{tibble} object.}

\item{extrm_vals}{A tibble (\code{tbl_df}) object with the following 5 fields:
\describe{
\item{\verb{Feature:}}{feature name matching fields in \code{data}}
\item{\verb{xtrm_min:}}{minimum acceptable value for that feature}
\item{\verb{xtrm_max:}}{maximum acceptable value for that feature}
\item{\verb{impute_min:}}{value to assign if below \code{xtrm_min}}
\item{\verb{impute_max:}}{value to assign if above \code{xtrm_max}}
}
Use \code{NA} to \emph{not} impute, or use \code{dplyr::filter()} to remove the entire
row if neither \code{min} nor \code{max} is desired for a given feature.}
}
\description{
Threshold RFU values according to the model-specific analyte QC ranges. One
form of imputation can be capping RFU values above a threshold (i.e.
winsorizing).
}
\examples{
x   <- data.frame(a = 1:3L, b = 4:6L, c = 7:9L, d = c(1.23, 4.56, 7.89))
tbl <- tibble::tribble(
  ~ Feature,  ~ xtrm_max, ~ impute_max, ~ xtrm_min, ~ impute_min,
    "a",         NA,        NA,           NA,         NA,
    "b",         5,         5,            0,          1,
    "c",         9,         7,            7.1,        7.1
)
impute_predictors(x, tbl)
}
\seealso{
Other impute: 
\code{\link{imputeNAs}()},
\code{\link{impute_outliers}()}
}
\author{
Stu Field
}
\concept{impute}
