% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeNAs.R
\name{imputeNAs}
\alias{imputeNAs}
\title{Impute NAs}
\usage{
imputeNAs(x)
}
\arguments{
\item{x}{A \code{soma_adat} object, a data frame, matrix,
or a numeric vector.}
}
\value{
An object of the same dimension and class as \code{x}
with any \code{NA} values replaced with the median of the
corresponding vector.
}
\description{
Imputes any \code{NAs}, usually in meta data, with the median value.
Must be a numeric. Objects of \code{factor} or \code{character} class
are not allowed.
}
\examples{
adat <- sim_test_data
imputeNAs(adat)   # 'soma_adat' S3

imputeNAs(adat$seq.3032.11)  # 'numeric' method; chosen at random

# test the S3 method
all.equal(imputeNAs(adat$seq.3032.11), imputeNAs(adat)$seq.3032.11)

x <- adat$seq.3032.11
x[seq(1, nrow(adat), 10)] <- NA_real_
x

median(x, na.rm = TRUE)

imputeNAs(x)

table(imputeNAs(x)) |> sort() |> tail()
}
\seealso{
\code{\link[=median]{median()}}, \code{\link[=modify_if]{modify_if()}}

Other impute: 
\code{\link{imputePredictors}()},
\code{\link{impute_outliers}()}
}
\author{
Stu Field
}
\concept{impute}
