% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{KFoldXVal}
\alias{KFoldXVal}
\title{K-fold Cross-Validation Samples}
\usage{
KFoldXVal(dat, k = 2, pct = NA, seed = NA)
}
\arguments{
\item{dat}{Data to split into cross-validation groups}

\item{k}{Constant k for k-fold cross-validation; defaults to 2}

\item{pct}{Percent of data to include in the first group; if not provided, defaults to equal-sized groups.
Remaining groups must be at least size 1, and are as equally sized as possible, with any remainder included
in the final group.}

\item{seed}{Optional random seed to use for sampling}
}
\description{
This function takes a dataframe or vector and returns a list of indices for use in cross-validation.
}

