% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance.R
\name{rebalance}
\alias{rebalance}
\title{Sampling for Class Imbalances}
\usage{
rebalance(data, var, method = c("down", "up"))
}
\arguments{
\item{data}{A data frame. For example, an individual
"analysis" fold (\code{\link[rsample:as.data.frame.rsplit]{rsample::analysis()}}) from an \code{\link[rsample:vfold_cv]{rsample::vfold_cv()}} call.}

\item{var}{Character. A column name of \code{data} containing class labels
for up-/down-sampling. Must contain a binary variable.
\code{tidyselect} variable selection is also supported. See Examples.}

\item{method}{Character (matched). One of "down" or "up". Down-sampling
\emph{decreases} the \code{n} per group to that of the smaller group size, whereas
up-sampling \emph{increases} (sampling with replacement) the \code{n} per group to
that of the larger group size. The default is "down".}
}
\value{
A sub- or re-sampled data frame that is smaller or larger than
\code{data} with equal class size.
}
\description{
Implements up- or down-sampling for analyses
with binary class imbalances.
}
\examples{
xtab <- globalr::cross_tab

# original class distribution
xtab(mtcars, vs)

# down-sampling `vs`
si <- rebalance(mtcars, "vs")    # 'down'
xtab(si, vs)

# up-sampling `vs` & passed variable
var <- "vs"
si <- rebalance(mtcars, var, "up")
xtab(si, vs)

# also supports unquoted strings
si  <- rebalance(mtcars, vs)
xtab(si, vs)
}
\author{
Stu Field
}
