% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove Statistical Outliers}
\usage{
remove_outliers(x, y = NULL, type = "nonparametric", ...)
}
\arguments{
\item{x}{\code{numeric(n)}. A vector of numeric values.}

\item{y}{Optional. If \code{NULL}, assume non-paired data and performs outlier
analysis on values in \code{x} along. If \emph{not} \code{NULL}, either a numeric vector
or character vector (e.g. class names) ordered in the same order as \code{x}
indicating the pairing.}

\item{type}{\code{character(1)}. Matched. Either
"parametric" or "nonparametric" to determine
the type of outliers detection implementation.}

\item{...}{Additional arguments passed to \code{\link[helpr:get_outliers]{helpr::get_outliers()}}.}
}
\value{
A \code{tibble} with columns \code{x} and \code{y} representing each
numeric vector pair with statistical outliers removed.
}
\description{
Remove statistical outliers from (optionally) \emph{paired} vectors of numeric
values using a statistical criterion based on median absolute deviation
(\eqn{6 \times mad}) and a fold change criterion (5x the median).
See "outlier detection" section for more information about
the \verb{type =} argument specification.
}
\section{outlier detection}{


There are 2 possible methods used to define an outlier measurement
and the return value depends on which method is implemented:
\enumerate{
\item The non-parametric case (default): agnostic to the distribution.
Outlier measurements are defined as falling outside \code{mad_crit * mad}
from the median \emph{and} a specified number of fold-changes from
the median (i.e. \code{fold_crit}; e.g. \eqn{5x}).\cr
\strong{Note:} \code{n_sigma} is ignored.
\item The parametric case: the mean and standard deviation are
calculated robustly via \code{\link[helpr:fit_gauss]{fit_gauss()}}. Outliers are defined
as measurements falling \emph{outside} +/- \code{n_sigma} * \eqn{\sigma}
from the the estimated \eqn{\mu}.\cr
\strong{Note:} \code{mad_crit} and \code{fold_crit} are ignored.
}

}

\examples{
x <- withr::with_seed(101, rnorm(10, mean = 1000, sd = 2))
x <- c(x, 10000)          # create outlier (11L)
x1 <- remove_outliers(x)  # 'x' only; no 'y'
x1

y  <- head(LETTERS, length(x))   # paired 'x' and 'y'
x2 <- remove_outliers(x, y)       # final row removed
x2
}
\seealso{
\code{\link[helpr:get_outliers]{helpr::get_outliers()}}
}
\author{
Stu Field
}
