% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-meta.R
\name{strip_meta}
\alias{strip_meta}
\title{Strip Meta Data}
\usage{
strip_meta(data)
}
\arguments{
\item{data}{A \code{soma_adat} object containing RFU feature data.}
}
\value{
A \code{data.matrix} object containing only the
data matrix of RFU values for the features.
}
\description{
Strips the (clinical) meta data from a \code{soma_adat} object
and returns a data matrix of only the RFU data.
}
\examples{
dim(sim_test_data)
class(sim_test_data)

strip_meta(sim_test_data) |> dim()
strip_meta(sim_test_data) |> class()
}
\seealso{
\code{\link[=data.matrix]{data.matrix()}}
}
\author{
Stu Field
}
