% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-meta.R
\name{strip_meta}
\alias{strip_meta}
\title{Strip Meta Data}
\usage{
strip_meta(data, feat = NULL)
}
\arguments{
\item{data}{A \code{data.frame} object containing
numeric feature data.}

\item{feat}{\code{charcter(n)}. A vector of column
names corresponding to the features.}
}
\value{
A numeric \code{data.matrix} object containing
only the data matrix of values for the features.
}
\description{
Strips the, non-feature (clinical), meta data from
a \code{data.frame} object and returns a data matrix
of only the feature (numeric) data.
}
\examples{
dim(sim_adat)
class(sim_adat)

strip_meta(sim_adat) |> dim()
strip_meta(sim_adat) |> class()
}
\seealso{
\code{\link[=data.matrix]{data.matrix()}}
}
\author{
Stu Field
}
