% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-date.R
\name{ToFiscalDate}
\alias{ToFiscalDate}
\title{Parse Fiscal Dates}
\usage{
ToFiscalDate(date, start.month = 7, output = "year")
}
\arguments{
\item{date}{Date to be parsed}

\item{start.month}{Starting month number of the fiscal year; defaults to 7 (July)}

\item{output}{Output format, including \code{year} for fiscal year number or \code{days}
for days passed since the beginning of the fiscal year}
}
\description{
This function takes a vector of dates and returns either the fiscal year, month,
or week, or months, weeks, or days since the start of the fiscal year.
}

